pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "db_tune_project"
version = "0.1.0"
description = "Autonomous DBA Optimization Agent"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.1.13",
    "streamlit"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "models*"]
[project.scripts]
server = "server.app:main"